+++ src/core/kv-storage/engines/browser.indexeddb/sync-engine.ts
@@ -1,0 +1,19 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+import { notImplement } from 'core/functools/not-implemented';
+
+export default [
+	'get',
+	'set',
+	'remove',
+	'keys'
+].reduce((engine, method) => {
+	engine[method] = notImplement({name: method}, () => undefined);
+	return engine;
+}, {});